src/requirements.txt
GitPython
requests
psutil
PySide6-Essentials>=6.8